requirements.txt
flask
gunicorn
numpy
pillow
scikit-learn
python-dotenv
onnxruntime